+++ mpstat.h
@@ -8,14 +8,14 @@
 
 
 struct mp_stats {
-   unsigned long cpu_idle			__attribute__ ((aligned (8)));
-   unsigned long cpu_iowait			__attribute__ ((packed));
-   unsigned long cpu_user			__attribute__ ((packed));
-   unsigned long cpu_nice			__attribute__ ((packed));
-   unsigned long cpu_system			__attribute__ ((packed));
-   unsigned long cpu_hardirq			__attribute__ ((packed));
-   unsigned long cpu_softirq			__attribute__ ((packed));
-   unsigned long irq				__attribute__ ((packed));
+   unsigned long long cpu_idle			__attribute__ ((aligned (8)));
+   unsigned long long cpu_iowait		__attribute__ ((packed));
+   unsigned long long cpu_user			__attribute__ ((packed));
+   unsigned long long cpu_nice			__attribute__ ((packed));
+   unsigned long long cpu_system		__attribute__ ((packed));
+   unsigned long long cpu_hardirq		__attribute__ ((packed));
+   unsigned long long cpu_softirq		__attribute__ ((packed));
+   unsigned long long irq			__attribute__ ((packed));
    /* Structure must be a multiple of 8 bytes, since we use an array of structures.
     * Each structure is *aligned*, and we want the structures to be packed together. */
 };
@@ -24,8 +24,8 @@ struct mp_stats {
 
 
 struct mp_timestamp {
-   unsigned long uptime;
-   unsigned long uptime0;
+   unsigned long long uptime;
+   unsigned long long uptime0;
    unsigned char hour;		/* (0-23) */
    unsigned char minute;	/* (0-59) */
    unsigned char second;	/* (0-59) */
